+++ src/server/app.js
@@ -1,16 +1,9 @@
 /* eslint-disable no-console */
-import React from 'react';
-import { Provider } from 'react-redux';
-import { renderToString } from 'react-dom/server';
-import { StaticRouter } from 'react-router';
-import { matchRoutes, renderRoutes } from 'react-router-config';
 import Raven from 'raven-js';
-
-import sc2 from 'sc2-sdk';
-import getStore from '../client/store';
-import routes from '../common/routes';
-import renderAmpPage, { compileAmpTemplate } from './renderers/ampRenderer';
-import renderSsrPage from './renderers/ssrRenderer';
+import { compileAmpTemplate } from './renderers/ampRenderer';
+import createSsrHandler from './handlers/createSsrHandler';
+import createAmpHandler from './handlers/createAmpHandler';
+import steemAPI from './steemAPI';
 
 const fs = require('fs');
 const express = require('express');
@@ -19,8 +12,6 @@ const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const http = require('http');
 const https = require('https');
-
-const steemAPI = require('./steemAPI');
 
 http.globalAgent.maxSockets = Infinity;
 https.globalAgent.maxSockets = Infinity;
@@ -56,56 +47,8 @@ const ampIndexPath = `${rootDir}/templates/amp_index.hbs`;
 const ampIndexHtml = fs.readFileSync(ampIndexPath, 'utf-8');
 const ampTemplate = compileAmpTemplate(ampIndexHtml);
 
-function serverSideResponse(req, res) {
-  const api = sc2.Initialize({
-    app: process.env.STEEMCONNECT_CLIENT_ID,
-    baseURL: process.env.STEEMCONNECT_HOST,
-    callbackURL: process.env.STEEMCONNECT_REDIRECT_URL,
-  });
-
-  if (req.cookies.access_token) {
-    api.setAccessToken(req.cookies.access_token);
-  }
-
-  const store = getStore(api);
-
-  const branch = matchRoutes(routes, req.url);
-  const promises = branch.map(({ route, match }) => {
-    const fetchData = route.component.fetchData;
-    if (fetchData instanceof Function) {
-      return fetchData(store, match);
-    }
-    return Promise.resolve(null);
-  });
-
-  return Promise.all(promises)
-    .then(() => {
-      const context = {};
-      const content = renderToString(
-        <Provider store={store}>
-          <StaticRouter location={req.url} context={context}>
-            {renderRoutes(routes)}
-          </StaticRouter>
-        </Provider>,
-      );
-      if (context.status) {
-        res.status(context.status);
-      }
-      res.send(renderSsrPage(store, content, indexHtml));
-    })
-    .catch((err) => {
-      Raven.captureException(err);
-      console.error('SSR error occured, falling back to bundled application instead', err);
-      res.send(indexHtml);
-    });
-}
-
-function ampResponse(req, res) {
-  steemAPI.sendAsync('get_content', [req.params.author, req.params.permlink]).then((result) => {
-    if (result.id === 0) res.sendStatus(404);
-    res.send(renderAmpPage(result, ampTemplate));
-  });
-}
+const ssrHandler = createSsrHandler(indexHtml);
+const ampHandler = createAmpHandler(ampTemplate);
 
 app.get('/callback', (req, res) => {
   const accessToken = req.query.access_token;
@@ -152,7 +95,7 @@ app.get('/i/:parent/@:referral/:permlink', (req, res) => {
     });
 });
 
-app.get('/:category/@:author/:permlink/amp', ampResponse);
-app.get('/*', serverSideResponse);
+app.get('/:category/@:author/:permlink/amp', ampHandler);
+app.get('/*', ssrHandler);
 
 module.exports = { app, server };
